Now items can be added one by one!
Now we've to use a loop to add the items one by one.

The previous committed code is printing the Array of items without any gap in first Item section. By using the MAP function we've fixed it.

See how the curElm and index is used here. If the key is not added then it throws an error.